backend/requirements/base.txt
Django==3.0.4
django-environ==0.4.5
djangorestframework==3.11.0
psycopg2==2.7.5
